Refactor: Move Hg raw diff code into the hg module
